Move routes to /routes dir
