+++ src/router/createRoutes.js
@@ -1,7 +1,8 @@
 import HomePage from "../components/HomePage";
-import FeaturedRoute from "./FeaturedRoute";
-import PhotoRoute from "./PhotoRoute";
-import NotFoundRoute from "./NotFoundRoute";
+
+import FeaturedRoute from "../routes/featured";
+import PhotoRoute from "../routes/photo";
+import NotFoundRoute from "../routes/not-found";
 
 export default function createRoutes() {
